moving around fixtures
